+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file. For commit 
 
   **Bugfixes**:
     - select and info bug after sorting https://github.com/gtsteffaniak/filebrowser/issues/277
+    - shares with public user
 
 ## v0.3.5